Add Hydra postbuild script
This postbuild script will add the build information json file and a
symbolic link to image to nix/store.

The names of the added items are printed on stdout, so they can be
scraped by hydrascraper from run command logs in hydra's web ui.

Link and json file are expected to be copied to cache in a cron job.

Signed-off-by: Tero Tervala <tero.tervala@unikie.com>