Add action.py as new scraper action script
This handles new way of providing build information. All info that is
available in the json file provided by hydra postbuild script via
nix/store is used from there in addition to information that is only
available via scraping from hydra.

If scraped information differs from info given in json file, then there
is a warning, and this can be easily changed to be an error if need be.

Signed-off-by: Tero Tervala <tero.tervala@unikie.com>