action.py: Return nonzero if no postbuild info
There is a time frame when Hydra UI has other build information, but
the postbuild script has not been run yet. After this change the
processing the build will be retried on the next scraper run.

Earlier, a build without runcommand logs was assumed to be old one
from the time before postbuild script, and thus it was just marked as
handled one.

By now all the old builds should be already in the handled builds
list, and we can assume all builds will have postbuild info eventually.

Signed-off-by: Tero Tervala <tero.tervala@unikie.com>