extracopy: Use actual file info from cache and web
Earlier "copied" file was causing trouble, so this version will get the
list of files from cache and web server instead of trying to keep up
it's own history of copied files.

If for some reason process is interrupted, on the next run the files
should still be copied normally.

cache and web copying can be enabled/disabled separately.

webcopy script was removed as all functionality is now in extracopy.sh.

Signed-off-by: Tero Tervala <tero.tervala@unikie.com>