Compress outputs for web upload
Output directories are compressed into a tar file.
Signatures are created for output directories and added into tar file.
Tar file is compressed with xz and uploaded to web server.

Creating the package happens in compress_outputs.sh which is called
from postbuild.py.

Uploading happens in "scheduled" webupload script. This is so that
if there are e.g. intermittent network issues, the files will
be uploaded once issues are cleared.

Upload script will keep uploaded files in "uploaded" directory for
four days before removing, this is mainly to debug any issues with
uploading files. Later this can be changed to stricter settings.

Package files are not added to cache, only uploaded to web server.

extracopy.sh was renamed cachecopy.sh and only copies signatures
to cache.

Signed-off-by: Tero Tervala <tero.tervala@unikie.com>